Move prepareColumnCache to StoragePython
